// src/modules/executor/components/EquipmentForm/types.ts
export interface EquipmentFormFormType {
  name: string;
  phone: string;
  email: string;
  type: string;
  craneType?: string;
  arrowLength?: string;
  tonnage?: string;
  groundType?: string;
  distance?: number;
  dumpTruckType?: string;
  bodyVolume?: number;
  addStuff?: string;
  ripper?: boolean;
  shoulderLength?: number;
  pumpVolume?: number;
}